Moved WorkflowServiceStubsImpl to implementation package. Added internal package-info
