+++ src/main/java/io/temporal/internal/package-info.java
@@ -1,0 +1,26 @@
+/*
+ *  Copyright (C) 2020 Temporal Technologies, Inc. All Rights Reserved.
+ *
+ *  Copyright 2012-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Modifications copyright (C) 2017 Uber Technologies, Inc.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ *  use this file except in compliance with the License. A copy of the License is
+ *  located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file. This file is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *  express or implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
+ */
+
+/**
+ * This package and its subpackages contain implementation classes of the Temporal SDK.
+ *
+ * <p><strong></strong>Do not use them directly in any application code! There is absolutely no
+ * guarantee about backwards compatibility of APIs in the internal packages.</strong>
+ */
+package io.temporal.internal;
